# server/docs/swagger.yaml
definitions:
  app.AppUsers:
    properties:
      authorityId:
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        type: string
      headerImg:
        type: string
      id:
        description: 主键ID
        type: integer
      nickName:
        type: string
      password:
        type: string
      phone:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /appUsers/createAppUsers:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建AppUsers
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.AppUsers'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建AppUsers
      tags:
      - AppUsers
  /appUsers/deleteAppUsers:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除AppUsers
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.AppUsers'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除AppUsers
      tags:
      - AppUsers
  /appUsers/deleteAppUsersByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除AppUsers
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除AppUsers
      tags:
      - AppUsers
  /appUsers/findAppUsers:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: authorityId
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: email
        type: string
      - in: query
        name: headerImg
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: nickName
        type: string
      - in: query
        name: password
        type: string
      - in: query
        name: phone
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: username
        type: string
      - in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询AppUsers
      tags:
      - AppUsers
  /appUsers/getAppUsersList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: authorityId
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: email
        type: string
      - in: query
        name: headerImg
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: nickName
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: password
        type: string
      - in: query
        name: phone
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: username
        type: string
      - in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取AppUsers列表
      tags:
      - AppUsers
  /appUsers/updateAppUsers:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新AppUsers
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.AppUsers'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新AppUsers
      tags:
      - AppUsers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"